# ARROW-2838: [Python] Speed up PandasObjectIsNull