Results for microbenchmarks.PandasObjectIsNull.time_PandasObjectIsNull:

* before:

```
               ========= =============
                  type
               --------- -------------
                  int     1.35±0.02ms
                 float    1.54±0.02ms
                 object     1.25±0ms
                decimal    20.8±0.5ms
               ========= =============
```

* after:

```
               ========= ============
                  type
               --------- ------------
                  int      811±10μs
                 float     1.54±0ms
                 object    1.37±0ms
                decimal   22.0±0.7ms
               ========= ============
```