@pitrou 2018-07-12
This speeds up the ARROW-2059 use case by 10-15%.